move tests for parser into single file
